moved lesion characterization as an integration test with test artifact
