--- experiments/utils.py
@@ -1,48 +1,0 @@
-import matplotlib.pyplot as plt
-
-# https://radiopaedia.org/articles/windowing-ct?lang=us
-display_settings = {
-    'brain': (80, 40),
-    'subdural': (300, 100),
-    'stroke': (40, 40),
-    'temporal bones': (2800, 600),
-    'soft tissues': (400, 50),
-    'lung': (1500, -600),
-    'liver': (150, 30),
-}
-
-
-def ctshow(img, window='soft tissues', fig=None, ax=None):
-    if fig is None or ax is None:
-        fig, ax = plt.subplots()
-    # Define some specific window settings here
-    if isinstance(window, str):
-        if window not in display_settings:
-            raise ValueError(f"{window} not in {display_settings}")
-        ww, wl = display_settings[window]
-    elif isinstance(window, tuple):
-        ww = window[0]
-        wl = window[1]
-    else:
-        ww = 6.0 * img.std()
-        wl = img.mean()
-
-    if img.ndim == 3:
-        img = img[0].copy()
-
-    ax.imshow(img, cmap='gray', vmin=wl-ww/2, vmax=wl+ww/2)
-    ax.set_xticks([])
-    ax.set_yticks([])
-    return ax.imshow(img, cmap='gray', vmin=wl-ww/2, vmax=wl+ww/2)
-
-
-def center_crop(img, thresh=-800, rows=True, cols=True):
-    cropped = img[img.mean(axis=1)>thresh, :]
-    cropped = cropped[:, img.mean(axis=0)>thresh]
-    return cropped
-
-
-def center_crop_like(img, ref, thresh=-800):
-    cropped = img[ref.mean(axis=1)>thresh, :]
-    cropped = cropped[:, ref.mean(axis=0)>thresh]
-    return cropped
